delete duplicated files
